hf_app/requirements.txt
gradio>=4.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=8.0.0
numpy>=1.21.0